analyzer/requirements.txt
# Database
psycopg2-binary==2.9.9

# LLM
anthropic==0.40.0

# HTTP Client
requests==2.31.0
urllib3==2.2.0

# Utilities
python-dotenv==1.0.1
python-dateutil==2.8.2
pytz==2024.1

# Retry Logic
tenacity==8.2.3

# Logging
colorlog==6.8.2
